refactor: ライブラリ関連の Model を移動
