+++ voicevox_engine/app/routers/library.py
@@ -7,8 +7,11 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException, Path, Request
 
 from voicevox_engine.engine_manifest import EngineManifest
-from voicevox_engine.library_manager import LibraryManager
-from voicevox_engine.model import DownloadableLibraryInfo, InstalledLibraryInfo
+from voicevox_engine.library_manager import (
+    DownloadableLibraryInfo,
+    InstalledLibraryInfo,
+    LibraryManager,
+)
 
 from ..dependencies import check_disabled_mutable_api
